====================================== BROKEN ATM =========================================== - support tag for PR #2273 - fixup the RulesClass function - fixup various file class - prorperly implement zip support for game load not yet loading assets - backport INIClass::Load handle the hooks internally - move around some hooks - replace PCX loading and mapping to more modern version

+++ src/Ext/BitFont/Body.cpp
@@ -41,7 +41,7 @@ int FakeBitFont::_Draw(const wchar_t* text, int xLeft, int yTop, int max_chars, 
 			{
 				ColorStruct rgb = Drawing::Int_To_RGB(saved_color);
 							rgb.Adjust(adjust_ratio, ColorStruct::Empty);
-				this->Color = rgb.ToInit();
+				this->Color = (WORD)rgb.ToInit();
 			}
 
 			adjust_ratio += 31;
